// src/enum/action.enum.tsx
export enum EActionEnum {
  GET_IS_LOADING = 'GET_IS_LOADING',
  SET_ACCESS_TOKEN = 'SET_ACCESS_TOKEN',
  GET_ALL_PROJECT = 'GET_ALL_PROJECT',
  GET_SEARCH_TEXT = 'GET_SEARCH_TEXT',
  GET_FILTER_STATUS = 'GET_FILTER_STATUS',
  SET_SHOW_DIALOG = 'SET_SHOW_DIALOG',
  SET_FORM_GENERAL = 'SET_FORM_GENERAL',
  SET_FORM_NOTIFICATION = 'SET_FORM_NOTIFICATION',
  NEXT_STEP_FORM = 'NEXT_STEP_FORM',
  PREV_STEP_FORM = 'PREV_STEP_FORM',
  SELECT_USER_MEMBER = 'SELECT_USER_MEMBER',
  SET_POSITION_USER = 'SET_POSITION_USER',
  REMOVE_USER_MEMBER = 'REMOVE_USER_MEMBER',
  SET_SELECT_ARRAY_USER = 'SET_SELECT_ARRAY_USER',
  SET_DATA_TASK = 'SET_DATA_TASK',
  REMOVE_TASK = 'REMOVE_TASK',
  ADD_TASK = 'ADD_TASK',
  UPDATE_TASK = 'UPDATE_TASK',
  PATCH_DATA_FORM = 'PATCH_DATA_FORM',
  PATCH_TEAM_DATA = 'PATCH_TEAM_DATA',
  PATCH_TASK_DATA = 'PATCH_TASK_DATA',
  SET_OPEN_MODAL_DIALOG = 'SET_OPEN_MODAL_DIALOG',
  SET_OPEN_MODAL_VIEW = 'SET_OPEN_MODAL_VIEW',
  SET_OPEN_CUSTOM_TIME = 'SET_OPEN_CUSTOM_TIME',
  SET_TITLE_ACTION_MODAL = 'SET_TITLE_ACTION_MODAL',
  GET_SELECT_PROJECT_ITEM = 'GET_SELECT_PROJECT_ITEM'
}